Move explore Lark rendering to presentation sink behind a thin facade
- reusable display implementation moves from capabilities/lark to
  loopx/presentation/sinks/lark/explore_results.py; capabilities/lark
  keeps a star-re-export facade so existing imports stay valid
- lark capability layout smoke covers the sink and facade equivalence
- explore docs state the role boundaries: explore capability owns
  structured exploration evidence (consumed by vision checkpoints,
  replanning, successor todos, and user gates first), presentation
  renders the public-safe projection, value connectors stay the
  external-signal/permission boundary
